+++ ClickHaskell/ClickHaskell/NativeProtocol/ServerPackets.hs
@@ -1,0 +1,74 @@
+{-# LANGUAGE
+    AllowAmbiguousTypes
+  , DataKinds
+  , DeriveAnyClass
+  , DeriveGeneric
+  , DerivingStrategies
+  , DuplicateRecordFields
+  , LambdaCase
+  , NamedFieldPuns
+  , OverloadedStrings
+  , UndecidableInstances
+#-}
+
+module ClickHaskell.NativeProtocol.ServerPackets where
+
+-- Internal dependencies
+import ClickHaskell.DbTypes
+import ClickHaskell.NativeProtocol.Serialization
+
+-- GHC included
+import GHC.Generics (Generic)
+
+-- * Server packets
+
+data ServerPacketType
+  = HelloResponse
+  | DataResponse
+  | Exception
+  | Progress
+  | Pong
+  | EndOfStream
+  | ProfileInfo
+  | Totals
+  | Extremes
+  | TablesStatusResponse
+  | Log
+  | TableColumns
+  | UUIDs
+  | ReadTaskRequest
+  | ProfileEvents
+  deriving (Show, Enum, Bounded)
+
+
+-- ** HelloResponse
+
+{-
+  https://github.com/ClickHouse/ClickHouse/blob/eb4a74d7412a1fcf52727cd8b00b365d6b9ed86c/src/Client/Connection.cpp#L520
+-}
+data HelloResponse = MkHelloResponse
+  { server_name                    :: ChString
+  , server_version_major           :: UVarInt
+  , server_version_minor           :: UVarInt
+  , server_revision                :: UVarInt
+  , server_parallel_replicas_proto :: UVarInt  `SinceRevision` DBMS_MIN_REVISION_WITH_VERSIONED_PARALLEL_REPLICAS_PROTOCOL
+  , server_timezone                :: ChString `SinceRevision` DBMS_MIN_REVISION_WITH_SERVER_TIMEZONE
+  , server_display_name            :: ChString `SinceRevision` DBMS_MIN_REVISION_WITH_SERVER_DISPLAY_NAME
+  , server_version_patch           :: UVarInt  `SinceRevision` DBMS_MIN_REVISION_WITH_VERSION_PATCH
+  , proto_send_chunked_srv         :: ChString `SinceRevision` DBMS_MIN_PROTOCOL_VERSION_WITH_CHUNKED_PACKETS
+  , proto_recv_chunked_srv         :: ChString `SinceRevision` DBMS_MIN_PROTOCOL_VERSION_WITH_CHUNKED_PACKETS
+  , password_complexity_rules      :: [PasswordComplexityRules] `SinceRevision` DBMS_MIN_PROTOCOL_VERSION_WITH_PASSWORD_COMPLEXITY_RULES
+  , read_nonce                     :: ChUInt64 `SinceRevision` DBMS_MIN_REVISION_WITH_INTERSERVER_SECRET_V2
+  }
+  deriving (Generic, Deserializable)
+
+instance Deserializable [PasswordComplexityRules] where
+  deserialize rev = do
+    _uvarInt <- deserialize @UVarInt rev
+    pure []
+
+data PasswordComplexityRules = MkPasswordComplexityRules
+  { original_pattern :: ChString
+  , exception_message :: ChString
+  }
+  deriving (Generic, Deserializable)
