--- ClickHaskell/ClickHaskell/NativeProtocol/Versioning.hs
@@ -1,115 +1,0 @@
-{-# LANGUAGE 
-    AllowAmbiguousTypes
-  , DataKinds
-  , DefaultSignatures
-  , DeriveAnyClass
-  , DeriveGeneric
-  , DerivingStrategies
-  , DuplicateRecordFields
-  , GeneralizedNewtypeDeriving
-  , LambdaCase
-  , NamedFieldPuns
-  , OverloadedStrings
-  , TemplateHaskell
-  , UndecidableInstances
-#-}
-
-module ClickHaskell.NativeProtocol.Versioning where
-
--- Internal dependencies
-import ClickHaskell.DbTypes
-import Paths_ClickHaskell (version)
-
--- GHC included
-import Data.Typeable (Proxy (..))
-import Data.Version (Version (..), showVersion)
-import GHC.TypeLits (KnownNat, Nat, natVal)
-import Language.Haskell.TH.Syntax (lift)
-
-
--- ** Client versioning
-
-clientMajorVersion, clientMinorVersion, clientPatchVersion :: UVarInt
-clientMajorVersion = fromIntegral $(lift (versionBranch version !! 0))
-clientMinorVersion = fromIntegral $(lift (versionBranch version !! 1))
-clientPatchVersion = fromIntegral $(lift (versionBranch version !! 2))
-
-clientNameAndVersion :: ChString
-clientNameAndVersion = $(lift ("ClickHaskell-" <> showVersion version))
-
-
--- ** Protocol compatibility wrappers
-
-type ProtocolRevision = UVarInt
-
-afterRevision
-  :: forall (revision :: Nat) monoid
-  .  (KnownNat revision, Monoid monoid)
-  => ProtocolRevision -> monoid -> monoid
-afterRevision chosenRevision monoid =
-  if chosenRevision >= (fromIntegral . natVal) (Proxy @revision)
-  then monoid
-  else mempty
-
-latestSupportedRevision :: ProtocolRevision
-latestSupportedRevision = (fromIntegral . natVal) (Proxy @DBMS_TCP_PROTOCOL_VERSION)
-
-data SinceRevision a (revisionNumber :: Nat) = MkSinceRevision a | NotPresented
-
-
--- ** Versions
-
-{-
-  Slightly modified C++ sources:
-  https://github.com/ClickHouse/ClickHouse/blob/eb4a74d7412a1fcf52727cd8b00b365d6b9ed86c/src/Core/ProtocolDefines.h#L6
--}
-type DBMS_TCP_PROTOCOL_VERSION = 54471;
-
-type DBMS_MIN_REVISION_WITH_CLIENT_INFO = 54032;
-type DBMS_MIN_REVISION_WITH_SERVER_TIMEZONE = 54058;
-type DBMS_MIN_REVISION_WITH_QUOTA_KEY_IN_CLIENT_INFO = 54060;
-type DBMS_MIN_REVISION_WITH_TABLES_STATUS = 54226;
-type DBMS_MIN_REVISION_WITH_TIME_ZONE_PARAMETER_IN_DATETIME_DATA_TYPE = 54337;
-type DBMS_MIN_REVISION_WITH_SERVER_DISPLAY_NAME = 54372;
-type DBMS_MIN_REVISION_WITH_VERSION_PATCH = 54401;
-type DBMS_MIN_REVISION_WITH_SERVER_LOGS = 54406;
-type DBMS_MIN_REVISION_WITH_CURRENT_AGGREGATION_VARIANT_SELECTION_METHOD = 54448;
-type DBMS_MIN_MAJOR_VERSION_WITH_CURRENT_AGGREGATION_VARIANT_SELECTION_METHOD = 21;
-type DBMS_MIN_MINOR_VERSION_WITH_CURRENT_AGGREGATION_VARIANT_SELECTION_METHOD = 4;
-type DBMS_MIN_REVISION_WITH_COLUMN_DEFAULTS_METADATA = 54410;
-type DBMS_MIN_REVISION_WITH_LOW_CARDINALITY_TYPE = 54405;
-type DBMS_MIN_REVISION_WITH_CLIENT_WRITE_INFO = 54420;
-type DBMS_MIN_REVISION_WITH_SETTINGS_SERIALIZED_AS_STRINGS = 54429;
-type DBMS_MIN_REVISION_WITH_SCALARS = 54429;
-type DBMS_MIN_REVISION_WITH_OPENTELEMETRY = 54442;
-type DBMS_MIN_REVISION_WITH_AGGREGATE_FUNCTIONS_VERSIONING = 54452;
-type DBMS_CLUSTER_PROCESSING_PROTOCOL_VERSION = 1;
-type DBMS_MIN_SUPPORTED_PARALLEL_REPLICAS_PROTOCOL_VERSION = 3;
-type DBMS_PARALLEL_REPLICAS_MIN_VERSION_WITH_MARK_SEGMENT_SIZE_FIELD = 4;
-type DBMS_PARALLEL_REPLICAS_PROTOCOL_VERSION = 4;
-type DBMS_MIN_REVISION_WITH_PARALLEL_REPLICAS = 54453;
-type DBMS_MERGE_TREE_PART_INFO_VERSION = 1;
-type DBMS_MIN_REVISION_WITH_INTERSERVER_SECRET = 54441;
-type DBMS_MIN_REVISION_WITH_X_FORWARDED_FOR_IN_CLIENT_INFO = 54443;
-type DBMS_MIN_REVISION_WITH_REFERER_IN_CLIENT_INFO = 54447;
-type DBMS_MIN_PROTOCOL_VERSION_WITH_DISTRIBUTED_DEPTH = 54448;
-type DBMS_MIN_PROTOCOL_VERSION_WITH_INCREMENTAL_PROFILE_EVENTS = 54451;
-type DBMS_MIN_REVISION_WITH_CUSTOM_SERIALIZATION = 54454;
-type DBMS_MIN_PROTOCOL_VERSION_WITH_INITIAL_QUERY_START_TIME = 54449;
-type DBMS_MIN_PROTOCOL_VERSION_WITH_PROFILE_EVENTS_IN_INSERT = 54456;
-type DBMS_MIN_PROTOCOL_VERSION_WITH_VIEW_IF_PERMITTED = 54457;
-type DBMS_MIN_PROTOCOL_VERSION_WITH_ADDENDUM = 54458;
-type DBMS_MIN_PROTOCOL_VERSION_WITH_QUOTA_KEY = 54458;
-type DBMS_MIN_PROTOCOL_VERSION_WITH_PARAMETERS = 54459;
-type DBMS_MIN_PROTOCOL_VERSION_WITH_SERVER_QUERY_TIME_IN_PROGRESS = 54460;
-type DBMS_MIN_PROTOCOL_VERSION_WITH_PASSWORD_COMPLEXITY_RULES = 54461;
-type DBMS_MIN_REVISION_WITH_INTERSERVER_SECRET_V2 = 54462;
-type DBMS_MIN_PROTOCOL_VERSION_WITH_TOTAL_BYTES_IN_PROGRESS = 54463;
-type DBMS_MIN_PROTOCOL_VERSION_WITH_TIMEZONE_UPDATES = 54464;
-type DBMS_MIN_REVISION_WITH_SPARSE_SERIALIZATION = 54465;
-type DBMS_MIN_REVISION_WITH_SSH_AUTHENTICATION = 54466;
-type DBMS_MIN_REVISION_WITH_TABLE_READ_ONLY_CHECK = 54467;
-type DBMS_MIN_REVISION_WITH_SYSTEM_KEYWORDS_TABLE = 54468;
-type DBMS_MIN_REVISION_WITH_ROWS_BEFORE_AGGREGATION = 54469;
-type DBMS_MIN_PROTOCOL_VERSION_WITH_CHUNKED_PACKETS = 54470;
-type DBMS_MIN_REVISION_WITH_VERSIONED_PARALLEL_REPLICAS_PROTOCOL = 54471;
